requirements.txt
django==1.11.17
djangorestframework==3.6.2
